added clang support, removed some warnings, openmp is made optional

+++ demos/LHC_restart.cpp
@@ -32,7 +32,7 @@ const double alpha = 1. / gamma_t / gamma_t; // First order mom. comp. factor
 
 // Tracking details
 uint N_t = 1000000;      // Number of turns to track; full ramp: 8700001
-int bl_target = 1.25e-9; // 4 sigma r.m.s. target bunch length in [s]
+double bl_target = 1.25e-9; // 4 sigma r.m.s. target bunch length in [s]
 
 int N_slices = 151;
 const std::string datafiles = DEMO_FILES "/LHC_restart/";
